🚀 feat(model): Enable Patchcore Training Half Precision (#3031)
patchcore infer dtype from model dtype

+++ src/anomalib/models/components/dimensionality_reduction/random_projection.py
@@ -193,4 +193,4 @@ class SparseRandomProjection:
             msg = "`fit()` has not been called on SparseRandomProjection yet."
             raise NotFittedError(msg)
 
-        return embedding @ self.sparse_random_matrix.T.float()
+        return embedding @ self.sparse_random_matrix.T.type(embedding.dtype)
